YaruCompactLayout: replace icon/titleBuilder with itemBuilder
And expose YaruNavigationRailItem as the recommended default building
block.
